+++ src/main/kotlin/dev/aaronhowser/mods/excessive_utilities/compatibility/jei/ModJeiPlugin.kt
@@ -107,6 +107,8 @@ class ModJeiPlugin : IModPlugin {
 
 		val disenchantmentRecipes = DisenchantmentFuelJeiCategory.getAllRecipes(level, registration)
 		registration.addRecipes(DISENCHANTMENT_FUELS, disenchantmentRecipes)
+
+		JeiInfoRecipes.addRecipes(registration)
 	}
 
 	override fun registerVanillaCategoryExtensions(registration: IVanillaCategoryExtensionRegistration) {