+++ src/main/kotlin/dev/aaronhowser/mods/excessive_utilities/datagen/tag/ModItemTagsProvider.kt
@@ -291,7 +291,6 @@ class ModItemTagsProvider(
 				ModBlocks.MAGIC_ABSORPTION.asItem(),
 				ModBlocks.MAGIC_INFUSER.asItem(),
 				ModBlocks.DEEP_DARK_PORTAL.asItem(),
-				ModBlocks.LAST_MILLENNIUM_PORTAL.asItem(),
 				ModBlocks.TRADING_POST.asItem(),
 				ModBlocks.INDEXER.asItem(),
 				ModBlocks.TRANSFER_FILTER.asItem(),