+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Added
 
 - Energy Transfer/Retrieval Nodes
+- The Last Millennium
 
 ### Changed